Adding checks before deleting previously generated files.

+++ convert.sh
@@ -3,11 +3,15 @@ dirs=$(find /files -name '*.drawio')
 
 for dir in $dirs
 do
-    echo "Converting ${dir} to ${dir}.pdf" 
-    rm $dir.pdf
+    echo "Converting ${dir} to ${dir}.pdf"
+    if test -f "${dir}.pdf"; then
+        rm $dir.pdf
+    fi
     node /home/node/draw.io-export/bin/drawio.js $dir -o $dir.pdf
     echo "Converting ${dir} to ${dir}.png" 
-    rm $dir.png
+    if test -f "${dir}.png"; then
+        rm $dir.png
+    fi
     node /home/node/draw.io-export/bin/drawio.js $dir -o $dir.png
 done
 echo "DRAW.IO-EXPORT Finished."
